Check docSize against max before mirroring (#45)
* Check docSize against max before mirroring

Prior to this commit, the MirroringUpdateProcessor had no check to
ensure that docs weren't running afoul of the batch size limit set at
the Kafka level.

This commit changes this to ensure that docs exceeding this limit are
not mirrored.  These offending docs may still be indexed, based on the
value of the URP's "indexUnmirrorableDocs" config property (which
defaults to 'false' if not set).