Check batch-size against "max" before mirroring (#48)
Prior to this commit, MirroringUpdateProcessor checked that individual
documents didn't exceed the mirroring batch-size limit on their own, but
no check existed to ensure that the entire list of docs doesn't exceed
the limit.

This commit adds that check (in MirroringUpdateProcessor.finish()).  If
the configured max batch-size is exceeded, MUP will log out an error
(which includes the IDs of all documents in the batch).