Ignore produer URP (if disabled) in standalone mode
XDC really only makes sense in a SolrCloud context.  But all the same,
its nice to handle MirroringUpdateRequestProcessorFactory more
gracefully if someone does accidentally configure their standalone core
to use the XDC producer URP.

This commit tweaks MirroringUpdateRequestProcessorFactory's
initialization so that in standalone mode it emits a warning about the
misconfiguration if disabled, and fails core init/load entirely if the
Mirroring URP is 'enabled=true'.