Add scripts to fetch+prepare solr-community data
One common barrier to running performance tests is the lack of a dataset
that's ready to be indexed into Solr.  Wikipedia dumps and other public
data is common, but there's work needed to massage them into a format
that's ready for Solr to index.

This commit partially addresses this problem by introducing scripts that
download a particular dataset and prepare it for indexing into Solr.
Hopefully this eliminates at least one of the hurdles in getting a quick
perf test running.

The dataset is comprised of the public interactions in our community.
Right now this covers interactions on our mailing lists, and data pulled
from Solr's git-log.  Hopefully in the future it can be supplemented
with data pulled from JIRA as well.  See
`scripts/community-dataset/README.md` for more details on the dataset
and how to use the provided scripts.