Add scripting to ease nightly benchmark running
Primary entrypoint is `run-benchmark-on-commits.sh`, which takes a
list of commits (or a "branch", "startCommit" and "endCommit")
and iterates over them, running Gatling benchmarks on each and storing
the results in a data directory.

The last piece of this needed to be truly usable is a cronjob-ready
script that can "remember" where the previous night's run left off, and
trigger benchmarking on any new commits since that.  That will be added
in a subsequent commit.