Add periodic-cronjob script and README on usage
`periodic-benchmark.sh` script isn't as nice as having a full Jenkins
setup, that's probably the "right" way to do this sort of recurring
automation.

But providing a simple path for folks to get up and running that doesn't
involve installing and running Jenkins seems like a nice 80/20 tradeoff
that lowers the barrier to entry a good deal.