Add automation for running gatling nightly (#124)
Primary entrypoint is `run-benchmark-on-commits.sh`, which takes a
list of commits (or a "branch", "startCommit" and "endCommit")
and iterates over them, running Gatling benchmarks on each and storing
the results in a data directory.

Also added is a `periodic-benchmark.sh` script that is a little more suitable
to run in a cronjob.  It works by recording some state about previous runs,
allowing the current run to "remember" where previous runs left off and
benchmark all commits since that point. 

A Jenkins integration or something else would obviously be nicer as a long
term solution.  That's probably the "right" way to do this sort of recurring
automation.

But providing a simple path for folks to get up and running that doesn't
involve installing and running Jenkins seems like a nice 80/20 tradeoff
that lowers the barrier to entry a good deal.

---------

Co-authored-by: Eric Pugh <epugh@opensourceconnections.com>